+++ vite.config.js
@@ -1,7 +1,7 @@
 import react from '@vitejs/plugin-react';
 import mdx from '@mdx-js/rollup';
 import remarkFrontmatter from 'remark-frontmatter';
-import remarkMdxFrontmatter from 'remark-mdx-frontmatter';
+import remarkMdxFrontmatter from './scripts/remark-mdx-frontmatter-local.mjs';
 import { defineConfig } from 'vite';
 import path from 'node:path';